refactored topbar (no longer connected to store) 6 added backbutton rendering
