--- src/components/TopBar/index.ts
@@ -1,10 +1,0 @@
-import { connect } from "react-redux";
-
-import TopBar from "./TopBar";
-import { StoreState } from "../../types/store";
-
-const mapStateToProps = (state: StoreState) => ({
-  amountLeft: state.amountLeft,
-});
-
-export default connect(mapStateToProps)(TopBar);
